% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{add_country_dates}
\alias{add_country_dates}
\title{add_country_dates}
\usage{
add_country_dates(countries_df)
}
\arguments{
\item{countries_df}{Optional.
If not provided, will use the package provided countries_data.
See get_countries for description of required input.}
}
\description{
(Legacy) Add dates to country data
}
\examples{
\dontrun{
countries_with_dates <- add_country_dates(countries_data)}
}
