% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskmatrix.R
\name{calc_gen_riskmatrix}
\alias{calc_gen_riskmatrix}
\title{calc_gen_riskmatrix}
\usage{
calc_gen_riskmatrix(df, population)
}
\arguments{
\item{df}{Dataframe with cumulative_cases and cumulative_deaths}

\item{population}{Population dataframe with population}
}
\description{
Calculate a risk matrix for a generalized df input.
}
\examples{
\dontrun{
riskmatrix_v3_df <- calc_riskmatrix_v3()}

}
