% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskmatrix.R
\name{calc_riskmatrix}
\alias{calc_riskmatrix}
\title{calc_riskmatrix}
\usage{
calc_riskmatrix(covid_sources, rts)
}
\description{
Calculate a risk matrix.
}
\examples{
\dontrun{
riskmatrix_df <- calc_riskmatrix()}
}
