% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskmatrix.R
\name{calc_riskmatrix_v2}
\alias{calc_riskmatrix_v2}
\title{calc_riskmatrix_v2}
\usage{
calc_riskmatrix_v2(covid_sources)
}
\description{
Calculate a risk matrix.
}
\examples{
\dontrun{
riskmatrix_v2_df <- calc_riskmatrix_v2()}

}
