% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskmatrix.R
\name{calc_riskmatrix_v3}
\alias{calc_riskmatrix_v3}
\title{calc_riskmatrix_v3}
\usage{
calc_riskmatrix_v3(df_ncov)
}
\description{
Caculate a risk matrix.
}
\examples{
\dontrun{
riskmatrix_v3_df <- calc_riskmatrix_v3()}

}
