% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rt.R
\name{calc_rts}
\alias{calc_rts}
\title{calc_rts}
\usage{
calc_rts(covid_sources)
}
\arguments{
\item{df_jhu}{Optional.
If not provided, will use the package provided countries_data and add_country_dates to make the most current dataset.
See get_countries for description of required input.}
}
\description{
Run EpiEstim on JHU and WHO.
}
