% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rt.R
\name{calc_who_rt}
\alias{calc_who_rt}
\title{calc_who_rt}
\usage{
calc_who_rt(df_who)
}
\arguments{
\item{df_jhu}{Optional.
If not provided, will use the package provided countries_data and add_country_dates to make the most current dataset.
See get_countries for description of required input.}
}
\description{
Run EpiEstim on WHO
}
