% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_whocriteria.R
\name{calc_whocriteria}
\alias{calc_whocriteria}
\title{calc_whocriteria}
\usage{
calc_whocriteria(covid_sources, df_testing_long, rts)
}
\arguments{
\item{country}{Optional.}
}
\description{
create binary variables for Mitigation re-opening for WHO criteria
}
