% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rt.R
\name{epi_curve_14}
\alias{epi_curve_14}
\title{epi_curve_14}
\usage{
epi_curve_14(country, ctyname)
}
\arguments{
\item{country}{Optional.
If not provided, will use the package provided countries_data and add_country_dates to make the most current dataset.
See get_countries for description of required input.}
}
\description{
Run EpiEstim.
}
