% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_googlemobility.R
\name{get_googlemobility}
\alias{get_googlemobility}
\title{get_googlemobility}
\usage{
get_googlemobility()
}
\description{
Get Google Mobility.
}
\examples{
\dontrun{
mobility_df <- get_googlemobility()}
}
