% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hotspot.R
\name{get_hotspots}
\alias{get_hotspots}
\title{get_hotspots}
\usage{
get_hotspots(df_ncov)
}
\description{
Hotspots for cumulative incidence or mortality. Need to supply JHU and WHO combined dataset.
Original code by Kathryn Winglee
Modified by Adam Vaughan & Michelle Schmitz / James Fuller / Imran Mujawar
May 1, 2020
This version adopts the script for Power BI
updated 9/24 with new algorithm from CBS GitHub page: https://github.com/cdcent/covid-response-analyses - Andrea Stewart
}
\examples{
\dontrun{
hotspot_df <- get_hotspot()}
}
