% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rt.R
\name{get_jhu_rt}
\alias{get_jhu_rt}
\title{get_jhu_rt}
\usage{
get_jhu_rt(df_jhu)
}
\arguments{
\item{df_jhu}{Optional.
If not provided, will use the package provided countries_data and add_country_dates to make the most current dataset.
See get_countries for description of required input.}
}
\description{
Run EpiEstim on JHU.
}
