% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onelist.R
\name{get_onelist}
\alias{get_onelist}
\title{get_onelist}
\usage{
get_onelist()
}
\description{
One list to rule them all and in keys bind them!
WB Income list <- Modified WHO COVID reporting list <- UN Pop Estimates
}
