% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rt.R
\name{get_rts}
\alias{get_rts}
\title{get_rts}
\usage{
get_rts(covid_sources)
}
\arguments{
\item{df_jhu}{Optional.
If not provided, will use the package provided countries_data and add_country_dates to make the most current dataset.
See get_countries for description of required input.}
}
\description{
Run EpiEstim on JHU and WHO.
}
