% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_sources.R
\name{get_who_covid}
\alias{get_who_covid}
\title{get_who_covid}
\usage{
get_who_covid(countries_dates)
}
\arguments{
\item{countries_dates}{Optional.
If not provided, will use the package provided countries_data and add_country_dates to make the most current dataset.
See get_countries for description of required input.}
}
\description{
Get and prepare WHO COVID data.
}
