% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hotspot.R
\name{hotspot}
\alias{hotspot}
\title{hotspot}
\usage{
hotspot(
  input.data,
  base.data,
  slope.cut,
  rate.cut,
  hb.cut,
  n,
  spar.param,
  dsource
)
}
\arguments{
\item{input.data}{get_covid_sources()}

\item{base.data}{case}

\item{slope.cut}{0.1}

\item{rate.cut}{10}

\item{hb.cut}{100}

\item{n}{7}

\item{spar.param}{0.6}

\item{dsource}{Either "WHO" or "JHU"}
}
\description{
Add dates to country data
}
