% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_maps.R
\name{map_template}
\alias{map_template}
\title{map_template}
\usage{
map_template(df, world, category_color_labels, category_color_values)
}
\arguments{
\item{df}{A dataframe with the following: region, country, date, factor value}

\item{world}{A dataframe with the following: geometry}

\item{category_color_labels}{List of labels that should map to the factor value of the df}

\item{category_color_values}{List of color values for mapping the labels. Needs to have the same length as category_color_labels!}
}
\description{
Cross-sectional map: Average daily incidence for the past 7 days for each country.
}
