% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_epicurve_double}
\alias{plot_epicurve_double}
\title{plot_epicurve_double}
\usage{
plot_epicurve_double(df)
}
\arguments{
\item{df}{A dataframe with the following: country, weekdate, cases and deaths}
}
\description{
(EPI WEEK) Visualize epi curve by cases and deaths.
Default viz for individual countries.
}
