% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_epicurve_double2}
\alias{plot_epicurve_double2}
\title{plot_epicurve_double2}
\usage{
plot_epicurve_double2(df)
}
\arguments{
\item{df}{A dataframe with the following: country, date, cases and deaths}
}
\description{
(DAILY) Visualize epi curve by cases and deaths.
Default viz for individual countries.
}
