% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_dashboard.R
\name{prod_itfdashboard}
\alias{prod_itfdashboard}
\title{prod_itfdashboard}
\usage{
prod_itfdashboard()
}
\description{
Produce CSVs needed for ITF dashboard
}
\examples{
\dontrun{
prod_itfdashboard()}

}
