% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_countryoverlays.R
\name{view_countryoverlays}
\alias{view_countryoverlays}
\title{view_countryoverlays}
\usage{
view_countryoverlays(covid_sources, googlemobility)
}
\description{
Create dataset for country overlays using covid_sources and Google mobility data.
}
